server/var/www/edc_app/requirements.txt
Babel==2.7.0
boto3==1.10.2
botocore==1.13.2
Django==2.2.8
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.8.0
django-formtools==2.1
django-otp==0.7.3
django-phonenumber-field==1.3.0
django-storages==1.7.2
django-tables2==2.1.1
django-taggit==1.1.0
django-two-factor-auth==1.9.1
djangorestframework==3.10.3
djangorestframework-jsonapi==3.0.0
docutils==0.15.2
inflection==0.3.1
jmespath==0.9.4
phonenumberslite==8.10.21
Pillow==6.2.1
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2019.3
qrcode==6.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
